+++ src/lib/pages/public-project/components/table/code/PublicProjectCodeTable.tsx
@@ -78,7 +78,7 @@ export const PublicProjectCodeTable = observer(
               variant="floating"
               value={searchKeyword}
               setInputState={setSearchKeyword}
-              label="Search with code ID or code name"
+              placeholder="Search with code ID or code name"
               size="md"
               mb={6}
             />