+++ src/lib/styles/theme/components/input.ts
@@ -27,7 +27,7 @@ export const Input: ComponentStyleConfig = {
         _invalid: {
           borderColor: "error.main",
         },
-        _placeholder: { color: "text.dark" },
+        _placeholder: { color: "pebble.600" },
       },
     },
   },